Splitting the controllers
