+++ activity_browser/application.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-import sys
 import os
 
 from PySide2.QtWidgets import QApplication
