+++ activity_browser/controllers/__init__.py
@@ -1,11 +1,12 @@
 # -*- coding: utf-8 -*-
-from .activity import ActivityController, ExchangeController
+from .activity import ActivityController
+from .exchange import ExchangeController
 from .database import DatabaseController
 from .parameter import ParameterController
-from .project import (
-    ProjectController, CSetupController, ImpactCategoryController
-)
-from .utils import UtilitiesController
+from .project import ProjectController
+from .impact_category import ImpactCategoryController
+from .calculation_setup import CalculationSetupController
+from .utilities import UtilitiesController
 from .plugin import PluginController
 
 controllers = {
@@ -15,7 +16,7 @@ controllers = {
     "parameter_controller": ParameterController,
     "plugin_controller": PluginController,
     "project_controller": ProjectController,
-    "cs_controller": CSetupController,
+    "cs_controller": CalculationSetupController,
     "ia_controller": ImpactCategoryController,
     "utils_controller": UtilitiesController,
 }
