+++ activity_browser/controllers/project.py
@@ -1,13 +1,11 @@
 # -*- coding: utf-8 -*-
-import traceback
-
 import brightway2 as bw
 from PySide2.QtCore import QObject, Slot
 from PySide2 import QtWidgets
 
 from activity_browser import log, signals, ab_settings
 from activity_browser.bwutils import commontasks as bc
-from activity_browser.ui.widgets import TupleNameDialog, ProjectDeletionDialog
+from activity_browser.ui.widgets import ProjectDeletionDialog
 
 class ProjectController(QObject):
     """The controller that handles all of the AB features on the level of
@@ -140,235 +138,3 @@ class ProjectController(QObject):
         # emit that the project list has changed because of the deletion,
         # regardless of a possible exception (which may have deleted the project anyways) 
         signals.projects_changed.emit()
-
-class CSetupController(QObject):
-    """The controller that handles brightway features related to
-    calculation setups.
-    """
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.window = parent
-
-        signals.new_calculation_setup.connect(self.new_calculation_setup)
-        signals.copy_calculation_setup.connect(self.copy_calculation_setup)
-        signals.rename_calculation_setup.connect(self.rename_calculation_setup)
-        signals.delete_calculation_setup.connect(self.delete_calculation_setup)
-
-    @Slot(name="createCalculationSetup")
-    def new_calculation_setup(self) -> None:
-        name, ok = QtWidgets.QInputDialog.getText(
-            self.window,
-            "Create new calculation setup",
-            "Name of new calculation setup:" + " " * 10
-        )
-        if ok and name:
-            if not self._can_use_cs_name(name):
-                return
-            bw.calculation_setups[name] = {'inv': [], 'ia': []}
-            signals.calculation_setup_selected.emit(name)
-            log.info("New calculation setup: {}".format(name))
-
-    @Slot(str, name="copyCalculationSetup")
-    def copy_calculation_setup(self, current: str) -> None:
-        new_name, ok = QtWidgets.QInputDialog.getText(
-            self.window,
-            "Copy '{}'".format(current),
-            "Name of the copied calculation setup:" + " " * 10
-        )
-        if ok and new_name:
-            if not self._can_use_cs_name(new_name):
-                return
-            bw.calculation_setups[new_name] = bw.calculation_setups[current].copy()
-            signals.calculation_setup_selected.emit(new_name)
-            log.info("Copied calculation setup {} as {}".format(current, new_name))
-
-    @Slot(str, name="deleteCalculationSetup")
-    def delete_calculation_setup(self, name: str) -> None:
-        # ask the user whether they are sure to delete the calculation setup
-        warning = QtWidgets.QMessageBox.warning(self.window,
-            f"Deleting Calculation Setup: {name}",
-            "Are you sure you want to delete this calculation setup?",
-            QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No,
-            QtWidgets.QMessageBox.No)
-        
-        # return if the users cancels
-        if warning == QtWidgets.QMessageBox.No: return
-
-        # otherwise try to delete the calculation setup
-        try:
-            del bw.calculation_setups[name]
-            signals.set_default_calculation_setup.emit()
-        # if an error occurs, notify the user and return
-        except Exception as e:
-            log.error(f"Deletion of calculation setup {name} failed with error {traceback.format_exc()}")
-            QtWidgets.QMessageBox.critical(self.window,
-                f"Deleting Calculation Setup: {name}",
-                "An error occured during the deletion of the calculation setup. Check the logs for more information",
-                QtWidgets.QMessageBox.Ok)
-            return
-
-        # inform the user that the calculation setup has been deleted
-        log.info(f"Deleted calculation setup: {name}")
-        QtWidgets.QMessageBox.information(self.window,
-            f"Deleting Calculation Setup: {name}",
-            "Calculation setup was succesfully deleted.",
-            QtWidgets.QMessageBox.Ok)
-
-    @Slot(str, name="renameCalculationSetup")
-    def rename_calculation_setup(self, current: str) -> None:
-        new_name, ok = QtWidgets.QInputDialog.getText(
-            self.window,
-            "Rename '{}'".format(current),
-            "New name of this calculation setup:" + " " * 10
-        )
-        if ok and new_name:
-            if not self._can_use_cs_name(new_name):
-                return
-            bw.calculation_setups[new_name] = bw.calculation_setups[current].copy()
-            del bw.calculation_setups[current]
-            signals.calculation_setup_selected.emit(new_name)
-            log.info("Renamed calculation setup from {} to {}".format(current, new_name))
-
-    def _can_use_cs_name(self, new_name: str) -> bool:
-        if new_name in bw.calculation_setups.keys():
-            QtWidgets.QMessageBox.warning(
-                self.window, "Not possible",
-                "A calculation setup with this name already exists."
-            )
-            return False
-        return True
-
-
-class ImpactCategoryController(QObject):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.window = parent
-
-        signals.copy_method.connect(self.copy_method)
-        signals.delete_method.connect(self.delete_method)
-        signals.edit_method_cf.connect(self.modify_method_with_cf)
-        signals.remove_cf_uncertainties.connect(self.remove_uncertainty)
-        signals.add_cf_method.connect(self.add_method_to_cf)
-        signals.delete_cf_method.connect(self.delete_method_from_cf)
-
-    @Slot(tuple, name="copyMethod")
-    def copy_method(self, method: tuple, level: str = None) -> None:
-        """Calls copy depending on the level, if level is 'leaf', or None,
-        then a single method is copied. Otherwise sets are used to identify
-        the appropriate methods"""
-        if level is not None and level != 'leaf':
-            methods = [bw.Method(mthd) for mthd in bw.methods if set(method).issubset(mthd)]
-        else:
-            methods = [bw.Method(method)]
-        dialog = TupleNameDialog.get_combined_name(
-            self.window, "Impact category name", "Combined name:", method, " - Copy"
-        )
-        if dialog.exec_() != TupleNameDialog.Accepted: return
-
-        new_name = dialog.result_tuple
-        for mthd in methods:
-            new_method = new_name + mthd.name[len(new_name):]
-            print('+', mthd)
-            if new_method in bw.methods:
-                warn = f"Impact Category with name '{new_method}' already exists!"
-                QtWidgets.QMessageBox.warning(self.window, "Copy failed", warn)
-                return
-            mthd.copy(new_method)
-            log.info("Copied method {} into {}".format(str(mthd.name), str(new_method)))
-        signals.new_method.emit()
-
-    @Slot(tuple, name="deleteMethod")
-    def delete_method(self, method_: tuple, level:str = None) -> None:
-        """Call delete on the (first) selected method and present confirmation dialog."""
-        if level is not None and level != 'leaf':
-            methods = [bw.Method(mthd) for mthd in bw.methods if set(method_).issubset(mthd)]
-        else:
-            methods = [bw.Method(method_)]
-        method = bw.Method(method_)
-        dialog = QtWidgets.QMessageBox()
-        dialog.setWindowTitle("Are you sure you want to delete this method?")
-        dialog.setText("You are about to PERMANENTLY delete the following Impact Category:\n("
-                       +", ".join(method.name)+
-                       ")\nAre you sure you want to continue?")
-        dialog.setIcon(QtWidgets.QMessageBox.Warning)
-        dialog.setStandardButtons(QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No)
-        dialog.setDefaultButton(QtWidgets.QMessageBox.No)
-        if dialog.exec_() == QtWidgets.QMessageBox.Yes:
-            for mthd in methods:
-                mthd.deregister()
-                log.info("Deleted method {}".format(str(mthd.name)))
-            signals.method_deleted.emit()
-
-    @Slot(list, tuple, name="removeCFUncertainty")
-    def remove_uncertainty(self, removed: list, method: tuple) -> None:
-        """Remove all uncertainty information from the selected CFs.
-
-        NOTE: Does not affect any selected CF that does not have uncertainty
-        information.
-        """
-        def unset(cf: tuple) -> tuple:
-            data = [*cf]
-            data[1] = data[1].get("amount")
-            return tuple(data)
-
-        method = bw.Method(method)
-        modified_cfs = (
-            unset(cf) for cf in removed if isinstance(cf[1], dict)
-        )
-        cfs = method.load()
-        for cf in modified_cfs:
-            idx = next(i for i, c in enumerate(cfs) if c[0] == cf[0])
-            cfs[idx] = cf
-        method.write(cfs)
-        signals.method_modified.emit(method.name)
-
-    @Slot(tuple, tuple, name="modifyMethodWithCf")
-    def modify_method_with_cf(self, cf: tuple, method: tuple) -> None:
-        """ Take the given CF tuple, add it to the method object stored in
-        `self.method` and call .write() & .process() to finalize.
-
-        NOTE: if the flow key matches one of the CFs in method, that CF
-        will be edited, if not, a new CF will be added to the method.
-        """
-        method = bw.Method(method)
-        cfs = method.load()
-        idx = next((i for i, c in enumerate(cfs) if c[0] == cf[0]), None)
-        if idx is None:
-            cfs.append(cf)
-        else:
-            cfs[idx] = cf
-        method.write(cfs)
-        signals.method_modified.emit(method.name)
-
-    @Slot(tuple, tuple, name="addMethodToCF")
-    def add_method_to_cf(self, cf: tuple, method: tuple):
-        method = bw.Method(method)
-        cfs = method.load()
-        # fill in default values for a new cf row
-        cfdata = (cf, {
-            'uncertainty type': 0,
-            'loc': float('nan'),
-            'scale': float('nan'),
-            'shape': float('nan'),
-            'minimum': float('nan'),
-            'maximum': float('nan'),
-            'negative': False,
-            'amount': 0
-        })
-        cfs.append(cfdata)
-        method.write(cfs)
-        signals.method_modified.emit(method.name)
-
-    @Slot(tuple, tuple, name="deleteMethodFromCF")
-    def delete_method_from_cf(self, to_delete: tuple, method: tuple):
-        method = bw.Method(method)
-        cfs = method.load()
-        delete_list = []
-        for i in cfs:
-            for d in to_delete:
-                if i[0][0] == d[0][0] and i[0][1] == d[0][1]:
-                    delete_list.append(i)
-        for d in delete_list:
-            cfs.remove(d)
-        method.write(cfs)
-        signals.method_modified.emit(method.name)
